Divide RowOption component for sub components
